Rework namespace functions and types
- Add set_namespaces function.
- Deprecate move_into_link_name_space and move_into_thread_name_spaces
  functions.
- Add NamespaceType.
- Deprecate ThreadNameSpaceType and LinkNameSpaceType.
- Add nsfs ioctls.
- Add NoArgGetter ioctl pattern.
